+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- generic: UninitializePlugin中にEditHandleを使った時のクラッシュを緩和する
 - module: `ReadSection`を`#[aviutl2::module::functions]`で受け取れるように
 - eframe: eguiのデフォルトフォントを同梱するように（feature flagによって制御できます）